RenderedTemplateFileHash strategy: append commit hash too
The added commit hash will not be used for change detection,
but it would be useful for finding the source file.

I.e. to get to the source file, one should navigate the appended
commit link and then browse to the source file in the tree
at that commit.

Even though the commit might not point to the last template HEAD,
if the source file didn't change its hash, then the file was not
changed between the written commit and the last HEAD, so both
hashes lead to the same file.